refactor: isolate integration catalog management
